Update home page to target employers
Just a starting point to flesh out the interal user experience.  A lot
of this could change once SFDL gets back to us with more specifics.

Change layout & text to flow with employer exp

This is simply a placeholder to have a starting point to develop the
other pages on the site.